requirements.txt
python-gitlab==4.4.0
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
httpx==0.25.0